fix bug command line 2
