+++ src/input.txt
@@ -1,0 +1,1001 @@
+1000
+7979
+7903
+6204
+3933
+3020
+2761
+4894
+536
+955
+7930
+5947
+1647
+5789
+7379
+9160
+3616
+2924
+4769
+39
+7138
+5425
+9688
+7079
+2389
+3595
+5941
+2234
+5596
+4591
+4399
+5914
+3496
+9776
+2272
+9496
+7795
+9586
+7395
+4503
+2432
+6258
+5979
+4840
+7446
+8963
+5042
+6766
+2980
+753
+6198
+60
+8404
+5824
+5175
+5652
+6093
+5386
+8897
+668
+7236
+5685
+9119
+6990
+4520
+4596
+6412
+4322
+8486
+5778
+3420
+4159
+7806
+2257
+9512
+2216
+4282
+5372
+2956
+6139
+2555
+6900
+6781
+9078
+9110
+6687
+1569
+5433
+2475
+5148
+5965
+4504
+3154
+1665
+7705
+320
+2707
+5684
+5179
+2348
+5223
+7491
+6101
+1207
+5995
+8613
+4386
+6596
+5531
+4288
+9609
+8545
+5666
+4251
+6026
+402
+1244
+6178
+1781
+1732
+8942
+2874
+6549
+4701
+1626
+4009
+2468
+8955
+189
+4711
+9502
+7820
+2744
+6826
+6565
+6089
+2817
+3169
+6233
+2805
+520
+1069
+3085
+9154
+9491
+1270
+4779
+8171
+5559
+9281
+4702
+2227
+2562
+9645
+4914
+2905
+8726
+6031
+1709
+7450
+6456
+9106
+213
+6676
+5074
+8377
+5413
+58
+6271
+3004
+3497
+5573
+3670
+9223
+2591
+5670
+1057
+7085
+7960
+8011
+328
+6465
+6948
+6019
+8342
+14
+6535
+6293
+2557
+112
+1101
+6226
+2588
+2920
+6618
+1132
+3764
+6016
+2609
+2886
+537
+6153
+1783
+5318
+8410
+9559
+9384
+1258
+805
+7326
+332
+9808
+1947
+9318
+4978
+259
+270
+2078
+9310
+6367
+9746
+5920
+1587
+6173
+9449
+1995
+8191
+7335
+3463
+2027
+702
+429
+4652
+8771
+2922
+8619
+2545
+4014
+2937
+8209
+8465
+9428
+1193
+9631
+1640
+4063
+9276
+4835
+5383
+5176
+9177
+7544
+3767
+1573
+5376
+8269
+4362
+797
+9157
+4279
+4546
+8934
+8074
+4533
+1778
+2639
+6332
+8011
+3310
+6169
+6675
+1712
+2564
+3559
+7685
+3796
+4865
+3294
+9247
+119
+2980
+4960
+4821
+791
+850
+4519
+4328
+8063
+1843
+1557
+3086
+2117
+6813
+5827
+5693
+906
+6565
+5060
+6105
+2873
+8731
+2200
+1876
+6709
+8450
+879
+7236
+6199
+2986
+1548
+7192
+7119
+3548
+1239
+2902
+4461
+2081
+7890
+4548
+9002
+1687
+2520
+5367
+4616
+1562
+4356
+319
+4945
+6244
+6129
+7505
+8825
+988
+2463
+4453
+5720
+9841
+1765
+4087
+3933
+8917
+4767
+1545
+9313
+7153
+422
+252
+3173
+3719
+1473
+1265
+7399
+5238
+3989
+2702
+8176
+4359
+4920
+333
+7383
+5222
+2242
+2853
+8077
+1467
+4006
+4027
+5954
+7365
+6321
+2362
+6797
+1393
+702
+383
+1296
+4091
+5491
+4421
+9031
+7292
+6610
+5958
+2888
+629
+5401
+7929
+7654
+8731
+82
+4605
+4478
+9359
+2416
+8244
+2729
+622
+5244
+9984
+7223
+7536
+7765
+6959
+3054
+6089
+9099
+4996
+4734
+9554
+8099
+9459
+7618
+5538
+5533
+7676
+4413
+8903
+457
+8550
+1799
+8211
+6438
+6473
+7108
+9713
+6575
+1782
+8054
+6788
+9583
+4611
+6137
+1565
+314
+5903
+3499
+9387
+2468
+292
+5050
+9263
+1473
+2492
+6911
+6629
+4243
+2708
+1899
+6120
+5645
+6280
+789
+5007
+5023
+4921
+5658
+9673
+252
+8177
+3174
+1040
+4456
+4415
+3423
+761
+9132
+6639
+4245
+4711
+3519
+1142
+9068
+6345
+79
+3871
+4149
+5843
+9171
+6280
+7496
+223
+9601
+6063
+5821
+8440
+6947
+3471
+218
+4365
+9568
+1971
+3237
+5069
+2716
+358
+7823
+7418
+3287
+5666
+2952
+4000
+324
+1958
+2123
+564
+2258
+3670
+6524
+6033
+7357
+3301
+5015
+3814
+2573
+6378
+6518
+6877
+8012
+1587
+9077
+7551
+7010
+1777
+859
+5964
+8163
+5312
+7535
+8386
+9972
+5062
+8609
+4295
+8022
+5269
+4222
+2228
+4989
+5444
+5452
+5843
+8398
+9423
+4185
+2618
+5751
+4349
+6075
+459
+529
+892
+7181
+4458
+5739
+7465
+878
+9017
+9062
+794
+785
+8700
+2727
+2146
+6589
+9740
+8662
+8839
+366
+6057
+2253
+8369
+9849
+8729
+3459
+5227
+3553
+4401
+8019
+8929
+280
+920
+7334
+6780
+9771
+1006
+4767
+6231
+8895
+7085
+5845
+595
+1126
+1346
+9451
+1057
+1666
+3018
+9755
+6998
+7113
+9601
+2057
+5440
+1854
+663
+1647
+9370
+5760
+1973
+6020
+5289
+205
+9460
+8319
+4150
+8171
+8647
+4319
+5077
+9014
+4261
+4965
+4846
+1411
+8656
+9851
+6044
+8394
+4612
+3548
+8429
+8360
+5548
+8394
+1282
+1222
+6839
+2840
+7950
+2996
+8728
+5191
+3221
+3513
+7879
+8843
+4988
+8226
+753
+5927
+438
+1348
+9806
+7088
+8576
+2281
+1933
+819
+5706
+4578
+5310
+3827
+5313
+3611
+5559
+2756
+3736
+3076
+8867
+1763
+5499
+7802
+3355
+5438
+8835
+4719
+2004
+9912
+4496
+7261
+689
+4915
+4443
+3545
+866
+1368
+117
+3549
+5626
+3997
+4108
+2870
+4799
+5082
+9692
+5853
+1801
+9667
+8572
+5052
+9014
+773
+6944
+7952
+9311
+4255
+1255
+955
+2835
+4562
+696
+2342
+7928
+2296
+5494
+9693
+6735
+2575
+2214
+8675
+3598
+1616
+2003
+3801
+3234
+5111
+6211
+9610
+4687
+5303
+4621
+9579
+3345
+8713
+3784
+1133
+6113
+442
+1156
+8024
+9735
+835
+1269
+5573
+3497
+302
+3268
+5592
+7850
+2772
+9437
+5549
+1973
+2902
+402
+1138
+2469
+7186
+5538
+8761
+5600
+2557
+3445
+5401
+8758
+2558
+5569
+1867
+2658
+2168
+4593
+2050
+3444
+9833
+6030
+7436
+5102
+3582
+7363
+1710
+2294
+6514
+6695
+8695
+9155
+9396
+1775
+5377
+4279
+5132
+398
+8105
+2691
+7606
+2675
+4849
+5428
+9482
+251
+4022
+9936
+7501
+5019
+3288
+9253
+8778
+182
+1459
+2188
+9996
+4863
+1169
+7666
+2149
+3433
+4089
+1173
+2968
+2553
+6359
+6505
+5762
+8360
+6625
+3298
+9805
+2892
+5810
+5396
+9758
+1141
+8535
+1864
+2570
+811
+7426
+9500
+76
+6116
+9195
+5229
+712
+7291
+9275
+3499
+747
+8710
+240
+4997
+2105
+340
+4135
+3070
+6463
+5337
+657
+6169
+3064
+8922
+3742
+4340
+8655
+5599
+4081
+854
+1064
+2019
+2923
+9691
+1704
+6550
+7870
+1301
+5413
+5617
+8443
+9242
+5337
+2264
+9305
+3199
+3752
+9167
+1832
+7144
+9723
+8728
+5174
+3928
+3794
+9321
+2123
+3272
+3773
+8485
+9690
+7155
+2839
+5512
+8292
+3559
+127
+830
+7367
+139
+1820
+5379
+7800
+6485
+4928
+9844
+940
+9571
+683
+5466
+6729
+535
+2197
+3165
+9711
+5269
+29
+2309
+8516
+5283
+3488
+2082
+5106
+3672
+2167
+1213
+79
+2353
+7748
+9712
+4611
+7065
+9323
+8563
+2857
+4237
+9776
+5057
+5559
+1096
+7035
+5991
+7240
+4623
+1869
+7625
+6132
+5031
+1994
+691
+7911
+1000
+3425
+613
+2975
+1672
+3236
+1534
+920
+1915
+4706
+8853
+7809
+6671
+2403
+3401
+4890
+9070
+3722
+9306
+6062
+5196
+859
+7286
+1447
+3363
+3421
+6297
+6854
+4198
+3092
+8177
+3908
+3462
+1842
+982
+7032
+4227
+8319
+254
+9367
+8994
+5667
+5283
+6966
+4167
